rm ParametricModel as interface was not clear (only *args), likelihood is not a ParametricModel (seperate module)
