+++ relife2/__init__.py
@@ -1,2 +1,3 @@
 from .fiability.distribution import Exponential, Gamma, Gompertz, LogLogistic, Weibull
+from .fiability.model import LifetimeModel, ParametricLifetimeModel, ParametricModel
 from .fiability.regression import AFT, ProportionalHazard
